+++ app/controls/AuthorSpan.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
     import type { RevAuthor } from "../messages/RevAuthor";
+    import { lastFocus } from "../stores";
     export let author: RevAuthor;
     export let includeTimestamp: boolean = false;
 
     let datetime = new Date(author.timestamp);
 
-    function relativeDate() {
+    function relativeDate(now: number) {
         const cutoffs = [60, 3600, 86400, 86400 * 7, 86400 * 30, 86400 * 365, Infinity];
         const units: Intl.RelativeTimeFormatUnit[] = ["second", "minute", "hour", "day", "week", "month", "year"];
         const formatter = new Intl.RelativeTimeFormat(undefined, {
@@ -13,7 +14,7 @@
             numeric: "auto",
         });
         let ticks = datetime.getTime();
-        let seconds = Math.round((ticks - Date.now()) / 1000);
+        let seconds = Math.round((ticks - now) / 1000);
         let unitIndex = cutoffs.findIndex((cutoff) => cutoff > Math.abs(seconds));
         let divisor = unitIndex ? cutoffs[unitIndex - 1] : 1;
 
@@ -28,7 +29,7 @@
             {author.name}
         </div>,
         <div class="inline" title={datetime.toLocaleString()}>
-            {relativeDate()}
+            {relativeDate($lastFocus)}
         </div>
     {:else}
         <div class="inline" title={author.email + ", " + datetime.toLocaleString()}>