+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Repository -> Init... and Repository -> Clone... menuitems, for creating repositories.
 - Progress bar for slow git operations (fetch, push, clone).
+- Relative timestamps update on each snapshot (which happen after modifications or when the window/tab is focused).
 - GG now respects the `snapshot.auto-update-stale` setting. Additionally, when first opening a repo, it will always update the working copy if it's stale.
 
 ### Fixed